+++ src/main/java/racingcar/utils/generator/MovableNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.utils;
+package racingcar.utils.generator;
 
 public class MovableNumberGenerator implements NumberGenerator {
     @Override
