+++ src/main/java/racingcar/utils/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.utils;
+package racingcar.utils.generator;
 
 public interface NumberGenerator {
     int generate();
