+++ src/main/java/racingcar/utils/generator/TrialNumberGenerator.java
@@ -1,4 +1,6 @@
-package racingcar.utils;
+package racingcar.utils.generator;
+
+import racingcar.utils.validator.TrialNumberValidator;
 
 public class TrialNumberGenerator {
     public static int generateTrialNumber(String input) {
