+++ src/test/java/racingcar/utils/NameValidatorTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import racingcar.utils.validator.NameValidator;
+
 class NameValidatorTest {
 
     @ParameterizedTest(name = "{index}: {1}")
