+++ src/test/java/racingcar/utils/TrialNumberGeneratorTest.java
@@ -2,8 +2,9 @@ package racingcar.utils;
 
 import static org.assertj.core.api.Assertions.*;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import racingcar.utils.generator.TrialNumberGenerator;
 
 public class TrialNumberGeneratorTest {
     @Test
