+++ src/test/java/racingcar/utils/TrialNumberValidatorTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import racingcar.utils.validator.TrialNumberValidator;
+
 public class TrialNumberValidatorTest {
     @ParameterizedTest(name = "{index}: {1}")
     @MethodSource("invalidParameters")
